Fix autocompletion bug
Fix a bug preventing the autocompletion of type-specific string
arguments for most widgets.

Use objectName and not proto.name as substring reference for object and
event properties retrieval, as objectName can get updated with a custom type
in the if-statement and would then not be equivalent to proto.name anymore.